Add source code, JUnit tests, and Ant buildfile.
The JPS source comprises three packages. The jps package contains the
main library class; jps.draw contains the implementation of the JPS
shape langauge; jps.postscript.commands contains representations of
PostScript commands. The source is almost unchanged from when my partner
and I submitted it during undergrad. I've cleaned up the code somewhat,
revised some of the existing Javadoc, and tweaked the package hierarchy,
but that's about it.

The class jpstests.JPSTests contains a few unit tests that exercise
various components of the shape language and PostScript translation.

The buildfile requires Ant 1.8 or better. It has targets for building,
testing, and cleaning up, as well as creating and testing a JAR file.