Add nonmarkup README, move Markdown one to .github
Although GitHub-hosted projects' READMEs are (I expect) primarily read
on the website, I insist on providing one that is reasonably legible in
raw form.  Unfortunately, I'm planning on changing the Markdown README
in a way that will make it significantly less legible.

Address this by creating a nonmarkup README that prioritizes legibility.
For now, it is an exact copy of the Markdown file, but more changes are
coming shortly.

(Since I'm already moving the Markdown file, might as well give it the
shorter ".md" extension.)